+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - biff.sqlite: add schema-module
 - biff.xtdb: add schema-module
 - biff.core: accept maps for :biff.core/init values
+- biff.tasks: format task adds blank lines to separate split form pairs
 
 **Breaking** [2.0.0-rc23](RELEASES/2.0.0-rc23.md) (2026-08-29):